# openapi/dungeonbuddy-statblocks-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelopeV1": {
        "additionalProperties": false,
        "description": "Top-level typed error body returned by v1 routes (never nested under ``detail``).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailV1"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelopeV1",
        "type": "object"
      },
...
      "GeneratedStatblockCandidateV1": {
        "additionalProperties": false,
        "properties": {
          "asset_brief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetBriefV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "asset_warnings": {
            "items": {
              "$ref": "#/components/schemas/AssetWarningV1"
            },
            "title": "Asset Warnings",
            "type": "array"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetRefV1"
            },
            "title": "Assets",
            "type": "array"
          },
          "candidate_id": {
            "pattern": "^cand_[a-z0-9]+$",
            "title": "Candidate Id",
            "type": "string"
          },
          "contract": {
            "const": "dungeonmind.dungeonbuddy-statblocks",
            "enum": [
              "dungeonmind.dungeonbuddy-statblocks"
            ],
            "title": "Contract",
            "type": "string"
          },
          "contract_version": {
            "const": "1.0.0",
            "enum": [
              "1.0.0"
            ],
            "title": "Contract Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "definition": {
            "$ref": "#/components/schemas/StatblockDefinitionV1-Output"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "generation_receipt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationReceiptV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_locator": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExactRevisionLocatorV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "validation_receipt": {
            "$ref": "#/components/schemas/ValidationReceiptV1"
          }
        },
        "required": [
          "candidate_id",
          "contract",
          "contract_version",
          "definition",
          "validation_receipt",
          "created_at",
          "expires_at"
        ],
        "title": "GeneratedStatblockCandidateV1",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidateDefinitionRequestV1": {
        "additionalProperties": false,
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/StatblockDefinitionV1-Input"
          }
        },
        "required": [
          "definition"
        ],
        "title": "ValidateDefinitionRequestV1",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "DungeonMindServer-owned DungeonBuddy statblock contract.",
    "title": "DungeonBuddy Statblocks v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/internal/dungeonbuddy/v1/statblock-candidates/{candidate_id}": {
      "get": {
        "operationId": "get_statblock_candidate_v1",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DungeonBuddy-Internal-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dungeonbuddy-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedStatblockCandidateV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Missing internal API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid internal API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Candidate not found"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Candidate expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Internal service misconfigured, provider unavailable, or persistence unavailable"
          }
        },
        "summary": "Get Candidate",
        "tags": [
          "DungeonBuddy Statblocks v1"
        ]
      }
    },
    "/api/internal/dungeonbuddy/v1/statblock-candidates:generate": {
      "post": {
        "operationId": "generate_statblock_candidate_v1",
        "parameters": [
          {
            "in": "header",
            "name": "X-DungeonBuddy-Internal-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dungeonbuddy-Internal-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCandidateRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedStatblockCandidateV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Missing internal API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid internal API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid request or generation validation failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Provider rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Unexpected generation failure (fail-closed unknown outcome)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Internal service misconfigured, provider unavailable, or persistence unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Provider timeout"
          }
        },
        "summary": "Generate Candidate",
        "tags": [
          "DungeonBuddy Statblocks v1"
        ]
      }
    },
    "/api/internal/dungeonbuddy/v1/statblock-candidates:revise": {
      "post": {
        "operationId": "revise_statblock_candidate_v1",
        "parameters": [
          {
            "in": "header",
            "name": "X-DungeonBuddy-Internal-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dungeonbuddy-Internal-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseCandidateRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedStatblockCandidateV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Missing internal API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid internal API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Source statblock or revision not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid request or generation validation failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Provider rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Unexpected generation failure (fail-closed unknown outcome)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Internal service misconfigured, provider unavailable, or persistence unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Provider timeout"
          }
        },
        "summary": "Revise Candidate",
        "tags": [
          "DungeonBuddy Statblocks v1"
        ]
      }
    },
    "/api/internal/dungeonbuddy/v1/statblock-definitions:validate": {
      "post": {
        "operationId": "validate_statblock_definition_v1",
        "parameters": [
          {
            "in": "header",
            "name": "X-DungeonBuddy-Internal-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dungeonbuddy-Internal-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDefinitionRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Missing internal API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid internal API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Internal service misconfigured, provider unavailable, or persistence unavailable"
          }
        },
        "summary": "Validate Definition Route",
        "tags": [
          "DungeonBuddy Statblocks v1"
        ]
      }
    },
    "/api/internal/dungeonbuddy/v1/statblocks": {
      "post": {
        "operationId": "create_statblock_v1",
        "parameters": [
          {
            "in": "header",
            "name": "X-DungeonBuddy-Internal-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dungeonbuddy-Internal-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStatblockRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateStatblockResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Missing internal API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid internal API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Statblock, revision, or acceptance candidate not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Idempotency conflict, parent mismatch, stale parent, or immutable ID conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Request validation or persistence-mode validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Internal service misconfigured, provider unavailable, or persistence unavailable"
          }
        },
        "summary": "Create Statblock",
        "tags": [
          "DungeonBuddy Statblocks v1"
        ]
      }
    },
...
    "/api/internal/dungeonbuddy/v1/statblocks/{statblock_id}": {
      "get": {
        "operationId": "get_statblock_v1",
        "parameters": [
          {
            "in": "path",
            "name": "statblock_id",
            "required": true,
            "schema": {
              "title": "Statblock Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DungeonBuddy-Internal-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dungeonbuddy-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatblockResourceV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Missing internal API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid internal API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Statblock or revision not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Internal service misconfigured, provider unavailable, or persistence unavailable"
          }
        },
        "summary": "Get Statblock",
        "tags": [
          "DungeonBuddy Statblocks v1"
        ]
      }
    },
...
    "/api/internal/dungeonbuddy/v1/statblocks/{statblock_id}/revisions/{revision_id}": {
      "get": {
        "description": "Resolve exactly the locator supplied by the caller; never select latest.",
        "operationId": "get_statblock_revision_v1",
        "parameters": [
          {
            "in": "path",
            "name": "statblock_id",
            "required": true,
            "schema": {
              "title": "Statblock Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DungeonBuddy-Internal-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dungeonbuddy-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatblockRevisionResourceV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Missing internal API key"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Invalid internal API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Statblock or revision not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeV1"
                }
              }
            },
            "description": "Internal service misconfigured, provider unavailable, or persistence unavailable"
          }
        },
        "summary": "Get Revision",
        "tags": [
          "DungeonBuddy Statblocks v1"
        ]
      }
    }
  }
}